Update blank.yml
I added maroon background, hyperlinks to the other pages, and it showing the recommendations were by you. I was not sure what your topic was so I just made the title and the name of the heading for the hyperlink "your title". Feel free to go through and replace it with your title name